+++ server/src/main/java/org/opensearch/gateway/remote/model/RemoteCustomMetadata.java
@@ -8,6 +8,7 @@
 
 package org.opensearch.gateway.remote.model;
 
+import org.opensearch.Version;
 import org.opensearch.cluster.metadata.Metadata.Custom;
 import org.opensearch.common.io.Streams;
 import org.opensearch.common.remote.AbstractClusterMetadataWriteableBlobEntity;
@@ -67,16 +68,17 @@ public class RemoteCustomMetadata extends AbstractClusterMetadataWriteableBlobEn
         final String customType,
         final String clusterUUID,
         final Compressor compressor,
-        final NamedWriteableRegistry namedWriteableRegistry
+        final NamedWriteableRegistry namedWriteableRegistry,
+        final Version version
     ) {
         super(clusterUUID, compressor, null);
         this.blobName = blobName;
         this.customType = customType;
         this.namedWriteableRegistry = namedWriteableRegistry;
-        this.customBlobStoreFormat = new ChecksumWritableBlobStoreFormat<>(
-            "custom",
-            is -> readFrom(is, namedWriteableRegistry, customType)
-        );
+        this.customBlobStoreFormat = new ChecksumWritableBlobStoreFormat<>("custom", is -> {
+            is.setVersion(version);
+            return readFrom(is, namedWriteableRegistry, customType);
+        });
     }
 
     @Override